Ziggy release 0.12.0 (2026-05-08)

+++ src/main/java/gov/nasa/ziggy/util/HumanReadableHeapSize.java
@@ -12,7 +12,7 @@ public class HumanReadableHeapSize {
         MB {
             @Override
             public float heapSizeGigabytes(float humanReadableHeapSize) {
-                return (int) humanReadableHeapSize / 1000;
+                return humanReadableHeapSize / 1000;
             }
         },
         GB {
